# app_authors/src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Documentación de la API de recursos de Authors
  description: Documento de los recursos de authors
  version: "1.0"

servers:
  - url: http://localhost:8088/q/openapi
    description: Download OPENAPI file

paths:
  /authors:
    get:
      responses:
        200:
          description: OK - Autores obtenidos exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        200:
          description: OK - Nuevo autor creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
  /authors/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          description: ID del autor en DB
          schema:
            type: integer
      responses:
        200:
          description: OK - autor obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
    put:
      parameters:
        - name: id
          in: path
          required: true
          description: ID del autor en DB
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        200:
          description: OK - autor actualizado exitosamente
          content:
            application/json: { }
    delete:
      parameters:
        - name: id
          in: path
          required: true
          description: ID del autor
          schema:
            type: integer
      responses:
        200:
          description: OK - autor eliminado exitosamente
          content:
            application/json: { }
components:
  schemas:
    Author:
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string